requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.28.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
apscheduler==3.10.4
jinja2==3.1.4
greenlet==3.1.1
python-multipart==0.0.12
aiofiles==24.1.0
pytz==2024.2
